pyproject.toml
[tool.poetry]
name = "vflip"
version = "1.0.1"
description = "Voltorb Flip engine and renderer"
authors = ["bijij <josh@josh-is.gay>"]
license = "MIT"
packages = [
    { include = "vflip" }
]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.0.0"

[tool.poetry.dev-dependencies]
black = "^21.0b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ["py38"]

[tool.pyright]
include = ["vflip/**"]
